Interact with localStorage in a more defined way
It turns out that the only reliable way to check for a key's existence
in localStorage is to check if .getItem(key) !== null. It also turns out
that despite how modern browsers have implemented it, localStorage was
designed to be able to hold arbitrary JavaScript objects, not just
strings, so scripts that assume that everything that is put into
localStorage is converted to a String are incorrect.

This commit makes sure this isn't one of those scripts.